# pkg/common/types.gen.go
// Package common provides primitives to interact the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen DO NOT EDIT.
package common

import (
	"time"
)

const (
	JwtAuthScopes = "jwtAuth.Scopes"
)

// CreateServerRequest defines model for CreateServerRequest.
type CreateServerRequest struct {
	Description string `json:"description"`
	Flavor      Flavor `json:"flavor"`
	Name        string `json:"name"`
	TemplateId  int    `json:"template_id"`

	// custom variables of this server
	TemplateVariables TemplateVariables `json:"template_variables"`
}

// Flavor defines model for Flavor.
type Flavor struct {

	// cpu size in micro cores (mcore), 1000mcore = 1core
	Cpu int `json:"cpu"`

	// memory size in MiB, 1024MiB = 1GiB
	Memory int `json:"memory"`

	// gpu unit
	NvidiaGpu int `json:"nvidia_gpu"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Id string `json:"id"`
	Pw string `json:"pw"`
}

// Quota defines model for Quota.
type Quota struct {

	// cpu size in micro cores (mcore), 1000mcore = 1core
	Cpu int `json:"cpu"`

	// instance count
	Instance int `json:"instance"`

	// memory size in MiB, 1024MiB = 1GiB
	Memory int `json:"memory"`

	// gpu unit
	NvidiaGpu int `json:"nvidia_gpu"`

	// storage size in GiB, 1024GiB = 1TiB
	Storage int `json:"storage"`
}

// Reason defines model for Reason.
type Reason struct {
	Reason string `json:"reason"`
}

// ServerObject defines model for ServerObject.
type ServerObject struct {

	// Creation date and time
	CreatedAt time.Time `json:"created_at"`

	// description of this server
	Description string `json:"description"`
	Flavor      Flavor `json:"flavor"`

	// id of this server
	Id int `json:"id"`

	// timestamp of last health check time
	LastProbeTime time.Time `json:"last_probe_time"`

	// timestamp of last transition of this server
	LastTransitionTime time.Time `json:"last_transition_time"`

	// show last event of this server. error message will be shown here.
	Message string `json:"message"`

	// name of this server
	Name  string   `json:"name"`
	Owner []string `json:"owner"`

	// show current status of this server
	Status   string   `json:"status"`
	Template Template `json:"template"`
}

// Template defines model for Template.
type Template struct {

	// kubernetes corev1/pod yaml based template
	Body        string `json:"body"`
	Description string `json:"description"`
	Id          int    `json:"id"`
	Name        string `json:"name"`

	// custom variables of this server
	Variables TemplateVariables `json:"variables"`
}

// TemplateVariable defines model for TemplateVariable.
type TemplateVariable struct {
	Name  string `json:"name"`
	Value string `json:"value"`
}

// TemplateVariables defines model for TemplateVariables.
type TemplateVariables []TemplateVariable

// UserInfo defines model for UserInfo.
type UserInfo struct {
	GroupQuota Quota  `json:"group_quota"`
	Id         string `json:"id"`
	UserQuota  Quota  `json:"user_quota"`
}

// AdminSetGroupQuotaJSONBody defines parameters for AdminSetGroupQuota.
type AdminSetGroupQuotaJSONBody Quota

// AdminSetUserQuotaJSONBody defines parameters for AdminSetUserQuota.
type AdminSetUserQuotaJSONBody Quota

// AdminCreateTemplateJSONBody defines parameters for AdminCreateTemplate.
type AdminCreateTemplateJSONBody Template

// LoginUserJSONBody defines parameters for LoginUser.
type LoginUserJSONBody LoginRequest

// CreateServerJSONBody defines parameters for CreateServer.
type CreateServerJSONBody CreateServerRequest

// UpdateUserInfoJSONBody defines parameters for UpdateUserInfo.
type UpdateUserInfoJSONBody struct {
	NewPassword *string `json:"new_password,omitempty"`
	Password    *string `json:"password,omitempty"`
}

// AdminSetGroupQuotaJSONRequestBody defines body for AdminSetGroupQuota for application/json ContentType.
type AdminSetGroupQuotaJSONRequestBody AdminSetGroupQuotaJSONBody

// AdminSetUserQuotaJSONRequestBody defines body for AdminSetUserQuota for application/json ContentType.
type AdminSetUserQuotaJSONRequestBody AdminSetUserQuotaJSONBody

// AdminCreateTemplateJSONRequestBody defines body for AdminCreateTemplate for application/json ContentType.
type AdminCreateTemplateJSONRequestBody AdminCreateTemplateJSONBody

// LoginUserJSONRequestBody defines body for LoginUser for application/json ContentType.
type LoginUserJSONRequestBody LoginUserJSONBody

// CreateServerJSONRequestBody defines body for CreateServer for application/json ContentType.
type CreateServerJSONRequestBody CreateServerJSONBody

// UpdateUserInfoJSONRequestBody defines body for UpdateUserInfo for application/json ContentType.
type UpdateUserInfoJSONRequestBody UpdateUserInfoJSONBody